extract checklist graphic components
